fix: 修正sigma参数的类型以确保正确创建Trackbar

+++ detector/LineDetect.py
@@ -117,7 +117,7 @@ class LineDetector(Detect):
         cv2.createTrackbar("maxLineGap", "Trackbar", self.maxLineGap, 600, self.__callback)
         cv2.createTrackbar("bias", "Trackbar", self.bias, 10, self.__callback)
         cv2.createTrackbar("odd_index", "Trackbar", self.odd_index, 20, self.__callback)
-        cv2.createTrackbar("sigma", "Trackbar", self.sigma * 10, 100, self.__callback)
+        cv2.createTrackbar("sigma", "Trackbar", int(self.sigma * 10), 100, self.__callback)
 
     def __callback(self, x):
         try:
